% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkinput.R
\name{check_input}
\alias{check_input}
\title{Check and Preprocess Input Data}
\usage{
check_input(input)
}
\arguments{
\item{input}{A data frame, matrix, vector, or factor variable to be converted
to a numeric matrix.}
}
\value{
A numeric matrix corresponding to the input data frame.
}
\description{
This function checks whether input data is a data frame, matrix, vector,
or factor variable, and converts it to a numeric matrix.
}
\keyword{internal}
