% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dftomat.R
\name{dftomat}
\alias{dftomat}
\title{Convert Data Frame to Numeric Matrix}
\usage{
dftomat(df)
}
\arguments{
\item{df}{A data frame to be converted to a numeric matrix.}
}
\value{
A numeric matrix corresponding to the input data frame.
}
\description{
This function converts a data frame into a numeric matrix.
If the data frame contains factor variables, they are first converted to
dummy variables (one-hot encoding).
If the data frame contains character variables, they are first converted to
factors and then to dummy variables.
}
\keyword{internal}
