% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnmodel.R
\name{nn_model}
\alias{nn_model}
\title{Define a Stochastic Generative Neural Network Model with Noise at Input Layer}
\usage{
nn_model(in_dim, noise_dim, hidden_dim = 100, out_dim, num_layer = 3)
}
\arguments{
\item{in_dim}{Integer. Input dimension.}

\item{noise_dim}{Integer. Dimension of the noise to inject.}

\item{hidden_dim}{Integer. Number of neurons in the hidden layers (default: 100).}

\item{out_dim}{Integer. Output dimension.}

\item{num_layer}{Integer. Number of layers in the model (default: 3).}
}
\value{
A generative neural network model with intermediate noise injection.
}
\description{
This function defines a generative neural network model for a certain architecture
and adds noise to the input layer.
}
\keyword{internal}
